refactor: massive cleaup
