+++ packages/fs-aws/tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "./build"
   },
   "include": ["src/"],
-  "references": [{ "path": "../source" }]
+  "references": [{ "path": "../source" }, {"path": "../fs"}]
 }
