+++ packages/fs/src/generator.ts
@@ -4,3 +4,9 @@ export async function toArray<T>(generator: AsyncGenerator<T>): Promise<T[]> {
   for await (const o of generator) output.push(o);
   return output;
 }
+
+/** Grab the first value from a async generator */
+export async function toFirst<T>(generator: AsyncGenerator<T>): Promise<T> {
+  const first = await generator.next();
+  return first.value;
+}
